requirements.txt
django==3.1.4
djangorestframework==3.12.2
mysqlclient==2.0.2
django-mysql==3.10.0
django-cors-headers==3.6.0